requirements.txt
keras==3.2.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
seaborn==0.13.2
tensorflow==2.16.1
tqdm==4.66.2
